+++ RELEASE_NOTES.md
@@ -41,6 +41,10 @@
 
 - Moved the auto-upload options to the upload modal (4.0.0)
 
+## _FFE_
+
+- Improved the upload modal (4.0.0)
+
 ### _ChessEvent_
 
 - Synchronize button moved to the _Data Transfer_ menu (4.0.0)